smv vv: update smokebot wrapper script

+++ Utilities/Firebot/smokebot_linux_wrapper.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
+if [ -e ~/SMOKEBOT/smokebot_running ] ; then
+  exit
+fi
 touch ~/SMOKEBOT/smokebot_running
 ~/SMOKEBOT/smokebot_linux.sh $*
 rm ~/SMOKEBOT/smokebot_running
